fix: rename ViCategoryContainer
